requirements.txt
pillow
qrcode[pil]
flet